Create user model with push tokens (#55)
Adds push tokens to the user model. I used the denormalized user model discussed over slack which I believe was part of #52 so I will coordinate with @bielleb. I made helper methods in the userController to make the user queryable by type (donor, volunteer, recipient, admin, or any). I also made a temporary "/api/tokens" endpoint to test getting all the user tokens, just so we can verify it is working.

This also adds a user router. I'm not sure if that is something we will need. I used it to create users for testing, but I will again discuss with Bernadette if this was something she was working on.