Add push token to user based on sub field of Auth0 JWT token (#70)
Uses sub field from auth0 jwt token to update notification token for specific user

Created a secure endpoint that accepts a new notification token and stores it to a user which has a unique "sub" field that can be decoded from any jwt token from auth0. These are unique to each user and guaranteed not to change, so it is more consistent than something like an email. I can work on getting persistent storage of the jwt tokens during sprint day so that task 25 from the umi-feeds-app repository can be fully resolved.